+++ mock/mocktest/path.go
@@ -1,8 +1,12 @@
 package mocktest
 
 import (
+	"context"
 	"errors"
 	"strings"
+	"testing"
+
+	"github.com/onsi/ginkgo/v2"
 )
 
 // A container for mock calls and a function for asserting results.
@@ -21,6 +25,43 @@ func (p path) Describe() string {
 	}
 	text = append(text, p.result.text)
 	return strings.Join(text, " > ")
+}
+
+// Run evaluates all declared mock client methods and assertions for the given test path.
+func (p path) Run(ctx context.Context, mck Mock) {
+	if mck.TestReporter == nil {
+		panic("Mock requires TestReporter, i.e. *testing.T, GinkgoT()")
+	}
+
+	for _, o := range p.once {
+		evalOnce(ctx, mck, o)
+	}
+	for _, c := range p.calls {
+		evalFn(ctx, mck, fn(c))
+	}
+
+	evalFn(ctx, mck, fn(p.result))
+}
+
+func evalFn(ctx context.Context, mck Mock, fun fn) {
+	switch tt := mck.TestReporter.(type) {
+	case *testing.T:
+		tt.Log(fun.text)
+	case ginkgo.GinkgoTInterface:
+		ginkgo.By(fun.text)
+	}
+
+	fun.does(ctx, mck)
+}
+
+func evalOnce(ctx context.Context, mck Mock, fun *once) {
+	if fun.ran {
+		return
+	}
+
+	evalFn(ctx, mck, fn(*fun))
+
+	fun.ran = true
 }
 
 type paths []path
